requirements.txt
Pillow
requests
selenium
bs4
chromedriver-autoinstaller